+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - On all external links is now a visible indicator !2280 @sefn
 - At stores under pickup history, storemanagers can now see the signed out foodsavers from pickup slots from past 7 days with the signout date. !2363 @martincodes-de
 - Voter Turnout Percentage can be seen if the poll has completed !2377 @viveknshah
+- Provide RestAPI to get backend defined common meta information for stores like "max count of slots per pickup" !1373 @koemai
 
 ## Changes
 - Checks in the id card creation function if only one foodsaver is selected and then reduces the